Problem: cargo was not available as part of the toolchain.
Solution: Ensure that cargo is built as part of the toolchain by enabling the
build to be extended (`--enable-extended`) and specifying cargo in the set of
tools to be built (`--tools=cargo`), both in build.sh.

Add a line in install.sh to copy the built cargo binary to the toolchain's
installation directory.