Problem: App Store rejects app containing Rust with bitcode error
There is a python program used by Apple (included inside the Xcode.app bundle)
which verifies a program's bitcode as part of rebuilding it. It includes a
check that each bitcode object has valid clang parameters included as part of
its XML xar header.

The Rust compiler sensibly chooses not to emit any clang parameters, since it
is not clang. Unfortunately this means any Rust-emitted bitcode will not pass
app store verification.

Locally, it produces an error like this one:

```
error: Clang option verification failed for bitcode 0038 (argument -emit-obj is
required)
```

Through App Connect, it produces an error like this one:

```
ITMS-90562: Invalid Bundle - The app cannot be processed because options not
allowed to be embedded in bitcode are detected in the submission
```

Solution: Patch Rust so that it will embed exactly the same clang cmdline that
Xcode/clang will when emitting Bitcode.

This patch is relatively brittle and not upstreamable in this form, but it
serves well to demonstrate the problem and is likely to be a workable solution
for now for anyone who wants to ship code.